backend/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
openai==1.13.3
